requirements.txt
absl-py==2.3.0
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
flatbuffers==25.2.10
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
narwhals==1.41.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
protobuf==5.29.5
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
